+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+* Add ability to configure stories runner class: `vividus-studio.stories-runner`
+
 ## [0.1.15] - 2024-06-01
 * Bump version of Gradle used to build projects from 8.7 to 8.8
 * Bump VS Code engine from 1.88.0 to 1.89.0